TurboScribeBot/requirements.txt
python-dotenv
selenium
requests